+++ ui/src/components/grid-generator/QasGridGenerator.vue
@@ -29,11 +29,6 @@ export default {
       type: [Array, Object, String]
     },
 
-    emptyResultValue: {
-      default: '-',
-      type: String
-    },
-
     headerClass: {
       default: 'text-bold',
       type: [Array, Object, String]
@@ -83,7 +78,7 @@ export default {
       const result = extend(true, {}, this.result)
       for (const key in result) {
         if (this.formattedFields[key]?.type) {
-          formattedResult[key] = humanize(this.formattedFields[key], result[key]) || this.emptyResultValue
+          formattedResult[key] = humanize(this.formattedFields[key], result[key]) || this.emptyResultText
           // TODO: Rever.
           // eslint-disable-next-line vue/no-side-effects-in-computed-properties
           this.slotValue[key] = { ...this.formattedFields[key], formattedResult: formattedResult[key] }