+++ ui/src/mixins/generator.js
@@ -7,6 +7,11 @@ export default {
       type: [Array, String, Object]
     },
 
+    emptyResultText: {
+      default: '-',
+      type: String
+    },
+
     fields: {
       default: () => ({}),
       type: Object